a. Display the output by applying arithmetic and logical operators on two numbers.Arithmetic operators perform their actions on numbers. Assignment operators assign values to variables. Logical operators compare two values and, based on whether the comparison is true (or false), return either a “true” or “false.”
Arithmetic operators:
Arithmetic Operators are the type of operators in C that are used to perform mathematical 
operations in a C program. They can be used in programs to define expressions and mathematical 
formulas.
Arithmetic Operators in C:
The C arithmetic operators are the symbols that are used to perform mathematical operations on 
operands. There are a total of 9 arithmetic operators in C to provide the basic arithmetic operations 
such as addition, subtraction, multiplication, etc.
Types of Arithmetic Operators in C:
The C Arithmetic Operators are of two types based on the number of operands they work. These are 
as follows:
1. Binary Arithmetic Operators
2. Unary Arithmetic Operators
1. Binary Arithmetic Operators in C
The C binary arithmetic operators operate or work on two operands. C provides 5 Binary 
Arithmetic Operators for performing arithmetic functions which are as follows:
Operator Name of the 
Operator Arithmetic Operation Syntax
[+Addition Add two operands. x + y.]
[–Subtraction Subtract the second operand from the first operand. x – y.]
[* Multiplication Multiply two operands. x * y.]
[/ Division Divide the first operand by the second operand. x / y.]
[% Modulus Calculate the remainder when the first operand is 
divided by the second operand. x % y]
2. Unary Arithmetic Operators in C
Programming Using ‘C’
General Engineering Department
The unary arithmetic operators operate or work with a single operand. In C, we have two unary 
arithmetic operators which are as follows:
Operator Symbol Operation Implementation
Decrement 
Operator --
Decreases the integer value of the variable 
by one. –h or h–
Increment 
Operator ++
Increases the integer value of the variable 
by one. ++h or h++